+++ src/constants/keys.js
@@ -1,5 +1,10 @@
 export const querykeys = {
   PROFILE: 'profile',
   INTRODUCE: 'introduce',
+  EDUCATION: 'education',
   CAREER: 'career',
+  PROJECT: 'project',
+  EXPERIENCE: 'experience',
+  SKILL: 'skill',
+  LINK: 'link',
 };
